+++ lib/stonk/money.rb
@@ -6,7 +6,7 @@ require "delegate"
 module Stonk
   class Money < SimpleDelegator
     def initialize(amount)
-      super(BigDecimal(amount))
+      super(BigDecimal(amount, Float::DIG + 1))
     end
   end
 end