+++ test/test_helper.rb
@@ -20,6 +20,7 @@ VCR.configure do |config|
   }
 
   config.filter_sensitive_data("<ALPHA_VANTAGE_TEST_API_KEY>") { ENV["ALPHA_VANTAGE_TEST_API_KEY"] }
+  config.filter_sensitive_data("<COIN_GECKO_TEST_API_KEY>") { ENV["COIN_GECKO_TEST_API_KEY"] }
 
   config.before_record(:alpha_vantage) do
     next unless ENV["ALPHA_VANTAGE_TEST_API_KEY"].nil?
@@ -29,6 +30,17 @@ VCR.configure do |config|
 
       In order to record cassettes against the Alpha Vantage API, you need
       to set the ALPHA_VANTAGE_TEST_API_KEY environment variable.
+    ERROR
+  end
+
+  config.before_record(:coin_gecko) do
+    next unless ENV["COIN_GECKO_TEST_API_KEY"].nil?
+
+    raise <<~ERROR
+      Error ENV['COIN_GECKO_TEST_API_KEY'] is not configured.
+
+      In order to record cassettes against the Coin Gecko API, you need
+      to set the COIN_GECKO_TEST_API_KEY environment variable.
     ERROR
   end
 end